% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisdoc.R
\name{analysisdoc}
\alias{analysisdoc}
\title{For exporting output as doc or Txt file}
\usage{
analysisdoc(filename, design, model, data, rep, factA, factB, factC)
}
\arguments{
\item{filename}{Type as 'filename.doc' or 'filename.txt' if you want output as text file}

\item{design}{Design is experimental technique you used in the experiment. rcbd -->  for randomized complete block  design , rcbd2 --> for 2 factorial  rcbd , rcbd3 --> for 3 factorial rcbd, crd -- > for complete randmized design crd2 --> for 2 factorial crd, crd3 --> for 3 factorial crd, lsd --> for latin square design ,split -- > for split plot design, strip --> for strip plot design}

\item{model}{0 for no test 1 for LSD 2 for Duncan 3 for HSD test}

\item{data}{data frame name}

\item{rep}{replication factor. Block for split plot, if block is not avalable put 0}

\item{factA}{Treatment name incase of LSD , RCBD and CRD}

\item{factB}{Second factor Name otherwise put 0 if its absent}

\item{factC}{third factor Name otherwise put 0 if its absent}
}
\value{

}
\description{
For exporting output as doc or Txt file
}
