% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape.R
\name{area2code}
\alias{area2code}
\title{seek area from data areacode and returns the code}
\usage{
area2code(area)
}
\arguments{
\item{area}{a string looking for}
}
\value{
a code if the area is found, else returns NA
}
\description{
seek area from data areacode and returns the code
}

