% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape.R
\docType{data}
\name{areacode}
\alias{areacode}
\title{Korean administrative area name with code}
\format{A data frame with 16 rows and 3 columns
\describe{
  \item{code}{code, numeric}
  \item{name}{the full name of Korean administrative area}
  \item{name2}{the abbreviation of name which is more commonly used}
}}
\usage{
areacode
}
\description{
A dataset containing the code and the name of Korean administrative area
}
\keyword{datasets}

