% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovalidate.R
\name{autovalidate}
\alias{autovalidate}
\title{Write out an assortment of package metadata necessary to run autovalidate}
\usage{
autovalidate(pkg = ".", output_dir = ".", steps = default_steps())
}
\arguments{
\item{output_dir}{A directory into which autovalidate files should be written}

\item{steps}{A list of functions to step through to generate autovalidate
files}
}
\description{
Write out an assortment of package metadata necessary to run autovalidate
}

\details{
Some details content
}