% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avu_log.R
\name{avu_log}
\alias{avu_log}
\title{Log to the autovalidateutils log file}
\usage{
avu_log(x, output = avu_opt("output"), file = avu_opt("errors"), append = TRUE)
}
\arguments{
\item{x}{A string or condition to emit to the log file}

\item{output}{A parent directory for all non-relative output files. Defaults to the
current working directory. (defaults to \code{file.path(normalize_path(tempdir()), "autovalidateutils")} from option \code{"autovalidateutils.output"})}

\item{file}{a \code{\link[base]{connection}}, or a character string naming
    the file to write to.  If \code{""}, print to the standard output
    connection.

    When \code{\link[base]{.Platform}$OS.type != "windows"}, and it
    is \code{"|cmd"}, the output is piped to the command given
    by \file{cmd}.
  }

\item{append}{if \code{TRUE} the data \code{x} are appended to the
    connection.}
}
\description{
Log to the autovalidateutils log file
\cr \cr \cr 
Also log this one but with a new line
This table is also worht adding
\tabular{rrrrr}{
\strong{mpg} \tab \strong{cyl} \tab \strong{disp} \tab \strong{hp} \tab \strong{drat} \cr
21.0 \tab 6 \tab 160 \tab 110 \tab 3.90\cr
21.0 \tab 6 \tab 160 \tab 110 \tab 3.90\cr
22.8 \tab 4 \tab 108 \tab  93 \tab 3.85\cr
21.4 \tab 6 \tab 258 \tab 110 \tab 3.08\cr
18.7 \tab 8 \tab 360 \tab 175 \tab 3.15
}
}
