% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{barn_en}
\alias{barn_en}
\title{Dataset with data from the children healthcare atlas, english version}
\description{
Dataset with data from the children healthcare atlas, english version
}
\keyword{datasets}
