\docType{data}
\name{bnames}
\alias{bnames}
\title{Popular baby names from 1880-2008}
\description{
  The top 1000 most popular boys names in the United States
  and the top 1000 most popular girls names in the United
  States for each year from 1880 - 2008
}
\details{
  The soundex variable is a code that describes the sund of
  the name. Names that sound similar will have the same
  soundex code.
}
\keyword{data}
\keyword{datasets}

