% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dagkir2}
\alias{dagkir2}
\title{Dataset with data from the day surgery healthcare atlas 2013-2017}
\description{
Dataset with data from the day surgery healthcare atlas 2013-2017
}
\keyword{datasets}
