% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dagkir2_eng}
\alias{dagkir2_eng}
\title{Dataset with data from the day surgery healthcare atlas 2013-2017, english version}
\description{
Dataset with data from the day surgery healthcare atlas 2013-2017, english version
}
\keyword{datasets}
