% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dagkir_en}
\alias{dagkir_en}
\title{Dataset with data from the day surgery healthcare atlas, english version}
\description{
Dataset with data from the day surgery healthcare atlas, english version
}
\keyword{datasets}
