% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_function.R
\name{data_function}
\alias{data_function}
\title{A data Function}
\usage{
data_function(love = TRUE)
}
\arguments{
\item{love}{Do you love data? Defaults to TRUE.}
}
\description{
This function allows you to express your love for data.
}
\examples{
data_function()
}
\keyword{data}
