% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{eldre_en}
\alias{eldre_en}
\title{Dataset with data from the elderly healthcare atlas, english version}
\description{
Dataset with data from the elderly healthcare atlas, english version
}
\keyword{datasets}
