% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example}
\alias{example}
\title{Example documentation}
\usage{
example(1, 2, 3)
}
\arguments{
\item{x}{value 1}

\item{y}{value 2}

\item{z}{value 3}
}
\description{
  Add values
}

\details{
Some details content
}
