% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstc_old.R
\name{get_msstc}
\alias{get_msstc}
\title{Get the MSSTC value for a given year}
\usage{
get_msstc(nc, year, lat = 63.1, lon = 165.5)
}
\arguments{
\item{nc}{A NetCDF file}

\item{year}{The year}

\item{lat}{Optional. The latitude to find data nearest to}

\item{lon}{Optional. The longitude to find data nearest to}
}
\description{
Get the MSSTC value for a given year
}
