% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstc.R
\name{get_msstc_data}
\alias{get_msstc_data}
\title{Get MSSTC data for a given year}
\usage{
get_msstc_data(year)
}
\arguments{
\item{year}{(numeric|character) The year to get data for}
}
\value{
(list) A NetCDF
}
\description{
https://psl.noaa.gov/data/gridded/data.ncep.reanalysis.surfaceflux.html
}
