% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountain-climbers.R
\name{get_team_info}
\alias{get_team_info}
\title{Extract team information from json string of team attributes}
\usage{
get_team_info(frame)
}
\description{
Extract team information from json string of team attributes
}
