% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstc.R
\name{get_time_indices}
\alias{get_time_indices}
\title{Get the indices for a given year}
\usage{
get_time_indices(nc, year)
}
\arguments{
\item{nc}{A NetCDF file}

\item{year}{The year to find an indices for}
}
\value{
(numeric) A vector of time indices
}
\description{
Get the indices for a given year
}
