% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gyn_en}
\alias{gyn_en}
\title{Dataset with data from the gynaecology healthcare atlas, english version}
\description{
Dataset with data from the gynaecology healthcare atlas, english version
}
\keyword{datasets}
