% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iajson2rdata.r
\name{iajson2rdata}
\alias{iajson2rdata}
\title{Read json-data from IA}
\usage{
iajson2rdata(json_file = NULL, rdata_file = NULL)
}
\arguments{
\item{json_file}{The json file used by IA}

\item{rdata_file}{The name of the RData file to be saved.
Will return an R object if NULL.
Default: NULL}
}
\value{
A data frame
}
\description{
All the data used to produce an Instant Atlas is stored in
a json file. This function will read such a file and put the
data into a data frame.
}
