% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape.R
\name{is.integer0}
\alias{is.integer0}
\title{Returns if x is integer(0)}
\usage{
is.integer0(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical value: TRUE for integer(0) otherwise FALSE
}
\description{
Returns if x is integer(0)
}

