% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape.R
\docType{data}
\name{korpopmap3}
\alias{korpopmap3}
\title{A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area, level 3
with merged population census data 2010}
\format{A Shape data with length of 3472}
\usage{
korpopmap3
}
\description{
A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area, level 3
with merged population census data 2010
}
\keyword{datasets}

