% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape.R
\name{mergeData2Map}
\alias{mergeData2Map}
\title{Merge data to an object of class SpatialPolygonsDataFrame}
\usage{
mergeData2Map(map, data, code = "code")
}
\arguments{
\item{map}{an object of class Shape(SpatialPolygonsDataFrame)}

\item{data}{a data.frame}

\item{code}{a name of variable of shp map data and data.frame to be matched with}
}
\value{
a Shape object with appended data
}
\description{
Merge data to an object of class SpatialPolygonsDataFrame
}

