% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstc_process.R
\name{msstc_process_single_file}
\alias{msstc_process_single_file}
\title{Process a single MSSTC raw file (daily values) and generate daily and monthly
derived files from it.}
\usage{
msstc_process_single_file(path, daily_path, monthly_path)
}
\description{
Process a single MSSTC raw file (daily values) and generate daily and monthly
derived files from it.
}
