% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSBM_pure.R
\name{NSBM.pure}
\alias{NSBM.pure}
\title{Nested species distribution modeling (pure bayes hierarchical)...}
\usage{
NSBM.pure(
  nsbm_obj,
  output = "probability",
  family = "binomial",
  link = "logit",
  prior.range = c(5, 0.01),
  prior.sigma = c(1, 0.01),
  nested_intercept = TRUE,
  auto_select_terms = FALSE,
  spatial = TRUE,
  mesh = NULL,
  n.threads = 1,
  cv.folds = 1,
  proj.new.env = TRUE,
  seed = NULL,
  save.output = FALSE
)
}
\arguments{
\item{nsbm_obj}{An object of class \code{nsdm.vinput}, result from \code{sabinaNSDM::NSDM.SelectCovariates()}.}

\item{output}{Character, \code{"intensity"} or \code{"probability"} (default).}

\item{family}{Character; \code{"binomial"} (default), supported \code{"poisson"}, \code{"nbinomial"}, or \code{"cp"}.}

\item{link}{Character; link function \code{"logit"} (default for \code{family = "binomial"}), otherwise \code{"log"}.}

\item{prior.range}{Numeric vector length 2. Prior on spatial range (e.g., \code{c(5, 0.01)}).}

\item{prior.sigma}{Numeric vector length 2. Prior on marginal standard deviation (e.g., \code{c(1, 0.01)}).}

\item{nested_intercept}{Logical; if TRUE = model regional intercept as deviation from global.}

\item{auto_select_terms}{Logical; if TRUE, select for each covariate automatically between “const”, “linear” or “rw2”.  #@@JMB no funciona bien aun}

\item{spatial}{Logical; include spatial latent field (SPDE) in the model (default: TRUE).}

\item{mesh}{An INLA mesh object created externally with \code{create_mesh()}. Required only if \code{spatial = TRUE}. Ignored if \code{spatial = FALSE}.}

\item{n.threads}{Number of threads for analysis}

\item{cv.folds}{Number of k-folds for cross-validation (default: 1 = no CV). If >1, returns mean ± sd AUC for global and regional models.}

\item{proj.new.env}{Logical. Whether to compute predictions under new scenarios (default: TRUE).}

\item{seed}{Optional integer to set the random seed for reproducibility.}

\item{save.output}{Logical. If TRUE, saves key model outputs (predictions, evaluation, summary ...).}
}
\value{
A named list of class \code{nsbm.inlabru} with the following elements:
\item{Species.Name}{Species name}
\item{args}{List of arguments used in the model fitting.}
\item{Selected.Variables.Global}{Names of selected global-scale covariates.}
\item{Selected.Variables.Regional}{Names of selected regional-scale covariates.}
\item{current.projections}{List with: prediction (\code{pred}) and spatial field (\code{pred_sp}).}
\item{new.projections}{List of projections to new.env (if \code{proj.new.env = TRUE}).}
\item{Summary}{\code{data.frame} with key evaluation metrics and significant variables.} #@@JMB revisar y refinar
}
\description{
Fits a nested species distribution model (NSDM) using hierarchical Bayesian modeling
with spatial structure via INLA and inlabru.
}
