% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nyfodt_en}
\alias{nyfodt_en}
\title{Dataset with data from the neonatal healthcare atlas, english version}
\description{
Dataset with data from the neonatal healthcare atlas, english version
}
\keyword{datasets}
