\name{OSDataAVS-package}
\alias{OSDataAVS-package}
\alias{OSDataAVS-package}
\docType{package}
\title{
Anvis outside evidence data repository
}
\description{
The repository contains comparables' data for further statistical analysis. The following group functions are covered:\cr
- contract R&D\cr
- contract Services\cr
- contract manufacturing\cr
}
\details{
\tabular{ll}{
Package: \tab AVSDAT\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2014-05-15\cr
License: \tab For Anvis use only\cr
}
As of now, the package provides the outside evidence data only. 
}
\author{
Dr. Steffen R. Möckel

Maintainer: Who to complain to <dr.s.moeckel@contrabass.biz>
}
\references{
Anvis Group Transfer Pricing Policy\cr
Comparables search strategy and Bureau van Dijks TP Catalyst searches
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
