% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountain-climbers.R
\name{recode_fitness_as_pct}
\alias{recode_fitness_as_pct}
\title{Recode arbitrary fitness value into a percentage of max.}
\usage{
recode_fitness_as_pct(frame)
}
\description{
Recode arbitrary fitness value into a percentage of max.
}
