% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape.R
\name{submap}
\alias{submap}
\title{Select a subdata of map}
\usage{
submap(map, area)
}
\arguments{
\item{map}{an object of class Shape(SpatialPolygonsDataFrame)}

\item{area}{a string of area looking for}
}
\value{
Subdata of class Shape of which code matched with area
}
\description{
Select a subdata of map
}

