\name{TF2mean_AddThis}
\alias{TF2mean}
\docType{data}
\title{
TF2mean_AddThis
}
\description{
Grid to go from tide free to mean tide. The grid should be added. TF2mean_AddThis is a 2880 x 1440 matrix.
}
\usage{load("TF2mean_AddThis.rda")}
\format{
 A matrix[lon,lat] of dimension 2880 x 1440 
}
\details{
...
}
\source{
...
}
\references{
Created by Ole B. Andersen \email{oa@space.dtu.dk}
}
\examples{
load("TF2mean_AddThis.rda")
}
\keyword{datasets}
